it2play: shutup int to float conversion warning
```
3rdparty/replay/it2play/it2drivers/hq.c:544:39: warning: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Wimplicit-const-int-float-conversion]
  544 |                 fPrng = (float)Random32() * (0.5f / INT32_MAX); // -0.5f .. 0.5f
      |                                                   ~ ^~~~~~~~~
/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/include/stdint.h:93:26: note: expanded from macro 'INT32_MAX'
   93 | #define INT32_MAX        2147483647
      |                          ^~~~~~~~~~
In file included from 3rdparty/replay/it2play/it2play_probe.cc:12:
3rdparty/replay/it2play/it2drivers/hq.c:552:39: warning: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Wimplicit-const-int-float-conversion]
  552 |                 fPrng = (float)Random32() * (0.5f / INT32_MAX); // -0.5f .. 0.5f
      |                                                   ~ ^~~~~~~~~
/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/include/stdint.h:93:26: note: expanded from macro 'INT32_MAX'
   93 | #define INT32_MAX        2147483647
```

+++ src/3rdparty/replay/it2play/it2drivers/hq.c
@@ -541,7 +541,7 @@ static int32_t HQ_PostMix(int16_t *AudioOut16, int32_t SamplesToOutput)
 	for (int32_t i = 0; i < SamplesTodo; i++)
 	{
 		// left channel - 1-bit triangular dithering
-		fPrng = (float)Random32() * (0.5f / INT32_MAX); // -0.5f .. 0.5f
+		fPrng = (float)Random32() * (0.5f / (float)INT32_MAX); // -0.5f .. 0.5f // mvtiaine: shutup int to float conversion warning
 		fOut = fMixBuffer[MixTransferOffset++] * MixGain;
 		fOut = (fOut + fPrng) - fPrngStateL;
 		fPrngStateL = fPrng;
@@ -549,7 +549,7 @@ static int32_t HQ_PostMix(int16_t *AudioOut16, int32_t SamplesToOutput)
 		*AudioOut16++ = (int16_t)CLAMP(out32, INT16_MIN, INT16_MAX);
 
 		// right channel - 1-bit triangular dithering
-		fPrng = (float)Random32() * (0.5f / INT32_MAX); // -0.5f .. 0.5f
+		fPrng = (float)Random32() * (0.5f / (float)INT32_MAX); // -0.5f .. 0.5f // mvtiaine: shutup int to float conversion warning
 		fOut = fMixBuffer[MixTransferOffset++] * MixGain;
 		fOut = (fOut + fPrng) - fPrngStateR;
 		fPrngStateR = fPrng;
